+++ packages/plugin-react-performance/lib/index.ts
@@ -1,1 +1,1 @@
-export { withInstrumentedComponent } from './plugin-react-performance'
+export { withInstrumentedComponent } from './with-instrumented-component';
