+++ packages/plugin-react-performance/lib/with-instrumented-component.tsx
