+++ packages/plugin-react-performance/tests/with-instrumented-component.test.tsx
